// src/modules/work-orders/domain/schemas/dto/response/process-work-order.response.ts
export interface ProcessWorkOrderResponse {
  action: string;
  recordId: string;
  workOrderId: string;
  orderCode?: string;
  cadastralKey?: string;
  previousStatus?: string;
  previousStatusName?: string;
  newStatus?: string;
  newStatusName?: string;
  currentStatus?: string;
  currentStatusName?: string;
  createdByUserId: string;
  processedByUserId?: string;
  approved?: boolean;
  comment?: string;
  processedAt: Date;
  metadata?: Record<string, string | number | boolean | null>;
}